修复删除部门BUG，新增供应商/单位接口

+++ src/main/java/com/example/demo/department/service/impl/DepartmentService.java
@@ -127,6 +127,9 @@ public class DepartmentService implements IDepartmentService {
 	 * @return
 	 */
 	public void deleteDepartmentsOfRecursion(List<Integer> idList) {
+		if(idList.isEmpty()) {
+			return;
+		}
 		// 删除父部门
 		departmentDao.deleteDepartments(idList);
 		// 查询子部门id
